优化 install.sh，简化备份和恢复 libsynosdk.so.7 的逻辑

+++ remotefs/install.sh
@@ -13,11 +13,9 @@ if [ "${1}" = "late" ]; then
 
   SO_FILE="/tmpRoot/usr/lib/libsynosdk.so.7"
   if [ -f "${SO_FILE}" ]; then
-    if [ ! -f "${SO_FILE}.bak" ]; then
-      echo "Backup ${SO_FILE}"
-      cp -vfp "${SO_FILE}" "${SO_FILE}.bak"
-    fi
     echo "Patching libsynosdk.so.7"
+    [ ! -f "${SO_FILE}.bak" ] && cp -f "${SO_FILE}" "${SO_FILE}.bak"
+    # force to support remote fs
     PatchELFSharp "${SO_FILE}" "SYNOFSIsRemoteFS" "B8 00 00 00 00 C3"
   else
     echo "libsynosdk.so.7 not found"
@@ -26,8 +24,5 @@ elif [ "${1}" = "uninstall" ]; then
   echo "Installing addon remotefs - ${1}"
 
   SO_FILE="/tmpRoot/usr/lib/libsynosdk.so.7"
-  if [ -f "${SO_FILE}.bak" ]; then
-    echo "Restore ${SO_FILE}"
-    mv -f "${SO_FILE}.bak" "${SO_FILE}"
-  fi
+  [ -f "${SO_FILE}.bak" ] && mv -f "${SO_FILE}.bak" "${SO_FILE}"
 fi
